+++ apps/api/src/app/asset/asset.module.ts
@@ -1,4 +1,4 @@
-import { AdminModule } from '@ghostfolio/api/app/admin/admin.module';
+import { AssetProfilesModule } from '@ghostfolio/api/app/endpoints/asset-profiles/asset-profiles.module';
 import { TransformDataSourceInRequestModule } from '@ghostfolio/api/interceptors/transform-data-source-in-request/transform-data-source-in-request.module';
 import { TransformDataSourceInResponseModule } from '@ghostfolio/api/interceptors/transform-data-source-in-response/transform-data-source-in-response.module';
 
@@ -9,7 +9,7 @@ import { AssetController } from './asset.controller';
 @Module({
   controllers: [AssetController],
   imports: [
-    AdminModule,
+    AssetProfilesModule,
     TransformDataSourceInRequestModule,
     TransformDataSourceInResponseModule
   ]