+++ libs/common/src/lib/interfaces/responses/asset-profile-response.interface.ts
@@ -2,7 +2,7 @@ import { MarketData } from '@prisma/client';
 
 import { EnhancedSymbolProfile } from '../enhanced-symbol-profile.interface';
 
-export interface MarketDataDetailsResponse {
+export interface AssetProfileResponse {
   assetProfile: Partial<EnhancedSymbolProfile>;
   marketData: MarketData[];
 }